requirements.txt

azure-functions==1.8.0
azure-identity==1.6.1
azure-kusto-data==2.3.2
azure-kusto-ingest==2.3.2
azure-storage-file-datalake==12.5.0
pandas==1.3.5
